% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Bayes_factors_consistency}
\alias{Bayes_factors_consistency}
\title{A dataset containing results for reproducing Figure 3 in Section 4 of the 
PGM 2018 and IJAR articles.}
\format{
A three-level list containing the Bayes factors obtained for the six 
scenarios described in Figures 3a-f. Each scenario has a list of nine matrices
with Bayes factors derived on increasing number of samples.
}
\usage{
Bayes_factors_consistency
}
\description{
A dataset containing results for reproducing Figure 3 in Section 4 of the 
PGM 2018 and IJAR articles.
}
\keyword{datasets}
