% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_BFCS_vectorized.R
\name{compute_BFCS}
\alias{compute_BFCS}
\title{Function for computing the Bayes Factors of Covariance Structures for a 
three-variable causal model.}
\usage{
compute_BFCS(corr, num_samples)
}
\arguments{
\item{corr}{Numeric 3x3 correlation matrix.}

\item{num_samples}{Integer number of samples.}
}
\value{
A vector containing the Bayes factor of all the eleven possible
conditional independence models on three variables relative to the fully
connected model.
}
\description{
Function for computing the Bayes Factors of Covariance Structures for a 
three-variable causal model.
}
\examples{
compute_BFCS(diag(3), 100)

}
