% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_BFCS_vectorized.R
\name{compute_BFCS_vectorized}
\alias{compute_BFCS_vectorized}
\title{Vectorized function for efficiently computing the Bayes Factors of many 
three-variable covariance structures at the same time.}
\usage{
compute_BFCS_vectorized(c12, c13, c23, num_samples)
}
\arguments{
\item{c12}{Vector of correlations between X_1 and X_2.}

\item{c13}{Vector of correlations between X_1 and X_3.}

\item{c23}{Vector of correlations between X_2 and X_3.}

\item{num_samples}{Integer number of samples.}
}
\value{
BFCS for all correlation combinations contained in the vectors
}
\description{
Vectorized function for efficiently computing the Bayes Factors of many 
three-variable covariance structures at the same time.
}
\examples{
cor_matrices <- apply(rWishart(100, 4, diag(3)), 3, cov2cor)
compute_BFCS_vectorized(
  c12 = cor_matrices[2, ], 
  c13 = cor_matrices[3, ], 
  c23 = cor_matrices[6, ], 
  num_samples = 1000
 )
}
