% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_BGe_score_vectorized.R
\name{compute_BGe_score_vectorized}
\alias{compute_BGe_score_vectorized}
\title{Vectorized compute Bayes factors of covariance structures using the Bayesian
Gaussian equivalent (BGe) score.}
\usage{
compute_BGe_score_vectorized(suffstat_list, num_samples, graph_list)
}
\arguments{
\item{suffstat_list}{List of sufficient statistics for a data set, containing
the means of the variables and their covariance matrix.}

\item{num_samples}{number of samples}

\item{graph_list}{List containing objects of class "bn", each of which
represents a graph structure for which we compute the BGe score.}
}
\value{
Matrix of size (number datasets) x (number graph structures) that
contains the BGe score for each combination of (dataset, graph).
}
\description{
Vectorized compute Bayes factors of covariance structures using the Bayesian
Gaussian equivalent (BGe) score.
}
\examples{
mec_list <- get_Markov_equivalence_class_list(format = "bnlearn")
suffstat_list = list(
  # three independent variables, zero mean
  list( 
    means = rep(0, 3),
    covmat = diag(3)
  ),
  # three independent variables, mean one
  list(
    means = rep(1, 3),
    covmat = diag(3)
  ),
  # a single variable repeated three times (correlation one)
  list(
    means = rep(0, 3),
    covmat = matrix(1, 3, 3)
  )
)
compute_BGe_score_vectorized(suffstat_list, num_samples = 10000, mec_list)
}
