% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_GRN.R
\name{compute_trigger_probabilities}
\alias{compute_trigger_probabilities}
\title{Compute the posterior probabilities of L_k -> T_i -> T_j for a simulated GRN
using the trigger package.}
\usage{
compute_trigger_probabilities(
  GRN,
  trigger_obj = NULL,
  seed = NULL,
  window.size = 50000
)
}
\arguments{
\item{GRN}{List produced by simulate_GRN containing GRN structure and statistics.}

\item{trigger_obj}{Precomputed object of class 'trigger' (build + loclink).}

\item{seed}{Random seed for reproducibility.}

\item{window.size}{Window size when looking for genetic markers. Default
option is value chosen in Chen et al. (2007)}
}
\value{
A numeric matrix of posterior probabilities for each possible causal
relationship between the expression traits.
}
\description{
Compute the posterior probabilities of L_k -> T_i -> T_j for a simulated GRN
using the trigger package.
}
\examples{
GRN <- simulate_GRN(100, 100, 100, 0.01)
compute_trigger_probabilities(GRN)
}
