% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_GRN.R
\name{derive_BFCS_posterior_probabilities}
\alias{derive_BFCS_posterior_probabilities}
\title{Compute the posterior probabilities of L_k -> T_i -> T_j for a simulated GRN 
using BFCS::compute_Bayes_factors_vectorized.}
\usage{
derive_BFCS_posterior_probabilities(
  GRN,
  prior = uniform_prior_GRN_DMAG(),
  loc_link = NULL
)
}
\arguments{
\item{GRN}{List produced by simulate_GRN containing GRN structure and statistics.}

\item{prior}{Prior on causal graphs (default uniform on DMAGs with bkg. info.)}

\item{loc_link}{Integer vector specifying genetic marker with strongest local
linkage for each expression trait. If not specified, we use the default BFCS 
maximum probability search strategy.}
}
\value{
A numeric matrix of posterior probabilities for each possible causal
relationship between the expression traits.
}
\description{
Compute the posterior probabilities of L_k -> T_i -> T_j for a simulated GRN 
using BFCS::compute_Bayes_factors_vectorized.
}
\examples{
GRN <- simulate_GRN(10, 10, 1000, 0.1)
derive_BFCS_posterior_probabilities(GRN)
derive_BFCS_posterior_probabilities(GRN, loc_link = 10:1)
}
