% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_GRN.R
\name{derive_BGe_posterior_probabilities}
\alias{derive_BGe_posterior_probabilities}
\title{Compute the posterior probabilities of L_k -> T_i -> T_j for a simulated GRN 
using bfcs::compute_BGe_score_vectorized.}
\usage{
derive_BGe_posterior_probabilities(GRN, prior = uniform_prior_GRN_DMAG())
}
\arguments{
\item{GRN}{List produced by simulate_GRN containing GRN structure and statistics.}

\item{prior}{Prior on causal graphs (default uniform on DMAGs with bkg. info.)}
}
\value{
A numeric matrix of posterior probabilities for each possible causal
relationship between the expression traits.
}
\description{
Compute the posterior probabilities of L_k -> T_i -> T_j for a simulated GRN 
using bfcs::compute_BGe_score_vectorized.
}
\examples{
GRN <- simulate_GRN(10, 10, 100, 0.1)
derive_BGe_posterior_probabilities(GRN)
}
