% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_Markov_equivalence_class_list}
\alias{get_Markov_equivalence_class_list}
\title{This function returns a list of all 11 different Markov equivalence classes
for causal graphs (DAGs) on three variables.}
\usage{
get_Markov_equivalence_class_list(format = "bnlearn")
}
\arguments{
\item{format}{The format in which to return the list. Currently, only "amat"
(adjacency matrix) and "bnlearn" (class "bn") formats are supported.}
}
\value{
A list containing the 11 different Markov equivalence classes.
}
\description{
This function returns a list of all 11 different Markov equivalence classes
for causal graphs (DAGs) on three variables.
}
\examples{
get_Markov_equivalence_class_list(format = 'bnlearn') # default
get_Markov_equivalence_class_list(format = 'amat')
}
