% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{off_diagonal}
\alias{off_diagonal}
\title{Function for extracting off-diagonal elements of a matrix.}
\usage{
off_diagonal(mat)
}
\arguments{
\item{mat}{Matrix of any type.}
}
\value{
Vector containing off-diagonal elements of matrix.
}
\description{
Function for extracting off-diagonal elements of a matrix.
}
\keyword{internal}
