% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_GRN.R
\name{simulate_GRN}
\alias{simulate_GRN}
\title{Generate gene regulatory network (GRN) data from structural equation model.}
\usage{
simulate_GRN(ngen, nexp, nobs, prob_edge, seed = NULL)
}
\arguments{
\item{ngen}{Integer number of genetic markers.}

\item{nexp}{Integer number of expression traits.}

\item{nobs}{Integer number of observations in the generated data set.}

\item{prob_edge}{Numeric probability of having a causal interaction (edge)
between two expression traits. Passed on to \link[pcalg]{randomDAG}.}

\item{seed}{Integer seed for reproducible random number generation.}
}
\value{
A data set consisting of \emph{nobs} samples from a random generated 
gene regulatory network. This includes \emph{ngen} binomially distributed 
variables representing the genetic markers associated with the genetic makeup 
of each individual observation. It also includes \emph{nexp} columns 
representing the gene expression levels of \emph{nexp} traits, which have an 
underlying causal graph describing their interactions.
}
\description{
Generate gene regulatory network (GRN) data from structural equation model.
}
\examples{
simulate_GRN(10, 10, 100, 0.1, 1634)
}
