% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_GRN_probabilities_IJAR}
\alias{simulated_GRN_probabilities_IJAR}
\title{A dataset containing results for reproducing Figures 4-7, as well as Table 2
from Section 4 in the IJAR article.}
\format{
A three-level list containing the computed posterior probabilities
for three different simulated GRNs (sparse, less dense, more dense) from
which (100, 1000) samples were generated, obtained using five different 
algorithms (trigger, BFCS DAG, BFCS DMAG, BFCS loclink, BGe).
}
\usage{
simulated_GRN_probabilities_IJAR
}
\description{
A dataset containing results for reproducing Figures 4-7, as well as Table 2
from Section 4 in the IJAR article.
}
\keyword{datasets}
