% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_GRN_probabilities_PGM}
\alias{simulated_GRN_probabilities_PGM}
\title{A dataset containing results for reproducing Figure 4, as well as Tables 2 
and 3 from Section 4 in the PGM 2018 article.}
\format{
A three-level list containing the computed posterior probabilities
for two different simulated GRNs (sparse, dense) from which (100, 1000) 
samples were generated, obtained using four different algorithms 
(trigger, BFCS DAG, BFCS DMAG, BGe).
}
\usage{
simulated_GRN_probabilities_PGM
}
\description{
A dataset containing results for reproducing Figure 4, as well as Tables 2 
and 3 from Section 4 in the PGM 2018 article.
}
\keyword{datasets}
