% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prior_structures.R
\name{uniform_prior_DAG_3var}
\alias{uniform_prior_DAG_3var}
\title{Uniform prior on three-variable DAG structures}
\usage{
uniform_prior_DAG_3var()
}
\value{
Logical vector indicating the probability of ...
}
\description{
Uniform prior on three-variable DAG structures
}
