% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prior_structures.R
\name{uniform_prior_GRN_DAG}
\alias{uniform_prior_GRN_DAG}
\title{Uniform prior on DAG structures (for GRN)}
\usage{
uniform_prior_GRN_DAG()
}
\description{
Uniform prior on DAG structures (for GRN)
}
