% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_BFCS_DMAG}
\alias{yeast_BFCS_DMAG}
\title{A dataset containing BFCS DMAG posterior probabilities on the yeast dataset
for reproducing Table 2 in Section 4 of the IJAR article, as well as Tables 1 
and 2 from Section 4 in the PGM 2018 article.}
\format{
6216x6216 matrix, where the element [i, j] indicates the estimated
probability that T_i -> T_j
}
\usage{
yeast_BFCS_DMAG
}
\description{
A dataset containing BFCS DMAG posterior probabilities on the yeast dataset
for reproducing Table 2 in Section 4 of the IJAR article, as well as Tables 1 
and 2 from Section 4 in the PGM 2018 article.
}
\keyword{datasets}
