% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_trigger_Chen}
\alias{yeast_trigger_Chen}
\title{A dataset containing posterior probabilities on the yeast dataset,
as reported in the supplement of [Chen et al. (2008)]
(https://genomebiology.biomedcentral.com/articles/10.1186/gb-2007-8-10-r219),
for reproducing Table 2 in Section 4 of the IJAR article, as well as Tables 1 
and 2 from Section 4 in the PGM 2018 article.}
\format{
6216x6216 matrix, where the element [i, j] indicates the estimated
probability that T_i -> T_j
}
\usage{
yeast_trigger_Chen
}
\description{
A dataset containing posterior probabilities on the yeast dataset,
as reported in the supplement of [Chen et al. (2008)]
(https://genomebiology.biomedcentral.com/articles/10.1186/gb-2007-8-10-r219),
for reproducing Table 2 in Section 4 of the IJAR article, as well as Tables 1 
and 2 from Section 4 in the PGM 2018 article.
}
\keyword{datasets}
