% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Bayes_factors.R
\name{compute_Bayes_factors}
\alias{compute_Bayes_factors}
\title{Bayes Factor Computation for a three-variable submodel.}
\usage{
compute_Bayes_factors(corr, no_samples)
}
\arguments{
\item{corr}{Input 3x3 correlation matrix}

\item{no_samples}{Number of samples}
}
\value{
A vector containing the evidence of various models
}
\description{
Bayes Factor Computation for a three-variable submodel.
}
