% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Bayes_factors.R
\name{compute_Bayes_factors_vectorized}
\alias{compute_Bayes_factors_vectorized}
\title{Vectorized compute Bayes factors of covariance structures.}
\usage{
compute_Bayes_factors_vectorized(c12, c13, c23, no_samples)
}
\arguments{
\item{c12}{vector of correlations between X_1 and X_2}

\item{c13}{vector of correlations between X_1 and X_3}

\item{c23}{vector of correlations between X_2 and X_3}

\item{no_samples}{number of samples}
}
\value{
Bayes factors for all correlations in the vectors
}
\description{
Vectorized compute Bayes factors of covariance structures.
}
