% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_parents_list}
\alias{get_parents_list}
\title{Small internal utility functions}
\usage{
get_parents_list(bn_graph)
}
\arguments{
\item{bn_graph}{object of class "bn" representing a bnlearn graph}
}
\value{
List of parents for each node, where the nodes are indexed from zero.
}
\description{
Small internal utility functions
}
\author{
Ioan Gabriel Bucur

Derive parents list from bnlearn graph for use in compute_BGe_score.
}
