% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prior_structures.R
\name{uniform_prior_GRN_DAG}
\alias{uniform_prior_GRN_DAG}
\title{Uniform prior on DAG structures with background knowledge.}
\usage{
uniform_prior_GRN_DAG()
}
\value{
Logical vector indicating the probability of Markov Equivalence
Classes over three variables, when X_1 precedes (X_2, X_3), assuming a
uniform prior over DAG causal structures.
}
\description{
Uniform prior on DAG structures with background knowledge.
}
\examples{
prior <- uniform_prior_GRN_DAG()
}
